+++ variables.tf
@@ -156,7 +156,6 @@ variable "enable_user_suffix" {
 variable "logs_destinations_ids" {
   type        = list(string)
   description = "List of destination resources Ids for logs diagnostics destination. Can be Storage Account, Log Analytics Workspace and Event Hub. No more than one of each can be set. Empty list to disable logging."
-  default     = null
 }
 
 variable "logs_categories" {