+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Unreleased
 
-Changed
+Breaking
   * AZ-160: Unify diagnostics settings on all Claranet modules
 
 # v3.1.1/v4.1.1 - 2021-03-09